SFD Business Test Data Example (#75)
* wip: working business example

* Add more examples

* Refactor: overall directory structure

* Refactor: personal and performance data

* Refactor business data

* Refactor business and performance data structures

* Remove accidental crn

* fix: broken personal tests

* Add: jsdoc comment

* Refactor: removed broken name examples

* Add: additional business examples for multiple invalid

* Refactor: add comments and helper functions

* Refactor: personal.js to reduce line length

* Fix: prettier lint

* Add: static organisation data overrides tests

* Improve comment

* Add: all invalid scenarios

* Add: more phone examples

* Refactor: use templates

* More refactoring
